Restructures console commands.
