+++ src/Tempest/Http/Console/RoutesCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tempest\Console\Commands;
+namespace Tempest\Http\Console;
 
 use Tempest\Console\Console;
 use Tempest\Console\ConsoleCommand;
