+++ src/Tempest/Http/IsResponse.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Tempest\Http;
 
+use function Tempest\get;
 use Tempest\Http\Cookie\Cookie;
 use Tempest\Http\Cookie\CookieManager;
 use Tempest\Http\Session\Session;
-use Tempest\View\View;
-use function Tempest\get;
 use function Tempest\view;
+use Tempest\View\View;
 
 trait IsResponse
 {
