+++ src/Tempest/Http/Mappers/RequestToObjectMapper.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Tempest\Http\Mappers;
 
 use Tempest\Http\Request;
-use Tempest\Mapper\Mapper;
 use function Tempest\map;
+use Tempest\Mapper\Mapper;
 
 final readonly class RequestToObjectMapper implements Mapper
 {
