+++ tests/Integration/Http/FileSessionTest.php
@@ -10,8 +10,8 @@ use Tempest\Http\Session\Session;
 use Tempest\Http\Session\SessionConfig;
 use Tempest\Http\Session\SessionId;
 use Tempest\Http\Session\SessionManager;
-use Tests\Tempest\Integration\FrameworkIntegrationTestCase;
 use function Tempest\path;
+use Tests\Tempest\Integration\FrameworkIntegrationTestCase;
 
 /**
  * @internal
