+++ tests/Integration/Mapper/MapperTest.php
@@ -8,6 +8,8 @@ use App\Modules\Books\Models\Author;
 use App\Modules\Books\Models\Book;
 use Tempest\Database\Migrations\CreateMigrationsTable;
 use Tempest\Database\Query;
+use function Tempest\make;
+use function Tempest\map;
 use Tempest\Mapper\Exceptions\MissingValuesException;
 use Tempest\Validation\Exceptions\ValidationException;
 use Tests\Tempest\Integration\FrameworkIntegrationTestCase;
@@ -17,8 +19,6 @@ use Tests\Tempest\Integration\Mapper\Fixtures\ObjectFactoryWithValidation;
 use Tests\Tempest\Integration\Mapper\Fixtures\ObjectWithBoolProp;
 use Tests\Tempest\Integration\Mapper\Fixtures\ObjectWithFloatProp;
 use Tests\Tempest\Integration\Mapper\Fixtures\ObjectWithIntProp;
-use function Tempest\make;
-use function Tempest\map;
 
 /**
  * @internal
