+++ tests/Integration/ORM/Mappers/QueryMapperTest.php
@@ -8,8 +8,8 @@ use App\Modules\Books\Models\Author;
 use App\Modules\Books\Models\Book;
 use Tempest\Database\Id;
 use Tempest\Database\Query;
-use Tests\Tempest\Integration\FrameworkIntegrationTestCase;
 use function Tempest\make;
+use Tests\Tempest\Integration\FrameworkIntegrationTestCase;
 
 /**
  * @internal
